Add SuaveCallBuilderExt + example
