+++ crates/network/src/lib.rs
@@ -6,3 +6,4 @@ mod contract;
 pub use network::SuaveNetwork;
 pub use signer::SuaveSigner;
 pub use provider::{SuaveProvider, SuaveFillProviderExt};
+pub use contract::SuaveCallBuilderExt;
